Moved the NodeUsage to NodeModel under cmd
