+++ main.go
@@ -3,25 +3,17 @@ package main
 import (
 	"flag"
 	"fmt"
+	"kubenodeusage/cmd/nodemodel"
 	"kubenodeusage/k8s"
 	"kubenodeusage/utils"
 	"os"
 	"reflect"
-	"regexp"
-	"sort"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/charmbracelet/bubbles/progress"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/iancoleman/strcase"
 )
-
-var helpStyle = lipgloss.NewStyle().Foreground(lipgloss.Color("#626262")).Render
 
 // map of keys of string type and values of interface type
 // Keys are strings.
@@ -65,289 +57,6 @@ func PrintArgs(args utils.Inputs) {
 
 }
 
-func DebugView(m model, output *strings.Builder) {
-	if m.args.Debug {
-		fmt.Fprint(output, " \nDebug mode enabled")
-		fmt.Fprint(output, "\nArgs: ", m.args)
-		fmt.Fprint(output, "\nNodes: ", m.nodestats)
-	}
-}
-
-func RightMetric(m model, index int) float32 {
-
-	switch m.args.Metrics {
-	case "memory":
-		if m.args.SortBy == "free" {
-			return float32(m.nodestats[index].Free_memory)
-		} else if m.args.SortBy == "capacity" || m.args.SortBy == "max" {
-			return float32(m.nodestats[index].Capacity_memory)
-		} else if m.args.SortBy == "color" || m.args.SortBy == "usage" {
-			return m.nodestats[index].Usage_memory_percent
-		}
-	case "cpu":
-		if m.args.SortBy == "free" {
-			return float32(m.nodestats[index].Free_cpu)
-		} else if m.args.SortBy == "capacity" || m.args.SortBy == "max" {
-			return float32(m.nodestats[index].Capacity_cpu)
-		} else if m.args.SortBy == "color" || m.args.SortBy == "usage" {
-			return m.nodestats[index].Usage_cpu_percent
-		}
-	case "disk":
-		if m.args.SortBy == "free" {
-			return float32(m.nodestats[index].Free_disk)
-		} else if m.args.SortBy == "capacity" || m.args.SortBy == "max" {
-			return float32(m.nodestats[index].Capacity_disk)
-		} else if m.args.SortBy == "color" || m.args.SortBy == "usage" {
-			return m.nodestats[index].Usage_disk_percent
-		}
-	}
-	// default return
-	return m.nodestats[index].Usage_memory_percent
-}
-
-func SortByHandler(m model) {
-
-	if m.args.SortBy != "" && m.args.SortBy != "name" && m.args.SortBy != "node" {
-		if !m.args.ReverseFlag {
-			sort.Slice(m.nodestats, func(i, j int) bool {
-				return RightMetric(m, i) < RightMetric(m, j)
-			})
-		} else {
-			sort.Slice(m.nodestats, func(i, j int) bool {
-				return RightMetric(m, i) > RightMetric(m, j)
-			})
-		}
-	} else if m.args.SortBy != "name" || m.args.SortBy != "node" {
-		if !m.args.ReverseFlag {
-			sort.Slice(m.nodestats, func(i, j int) bool {
-				return m.nodestats[i].Name < m.nodestats[j].Name
-			})
-		} else {
-			sort.Slice(m.nodestats, func(i, j int) bool {
-				return m.nodestats[i].Name > m.nodestats[j].Name
-			})
-		}
-	}
-
-}
-func ApplyFilters(m model) []k8s.Node {
-	if m.args.FilterLabel != "" {
-		return FilterForLabel(m)
-	} else if m.args.FilterNodes != "" {
-		return FilterForNode(m)
-	} else if m.args.FilterColor != "" {
-		return FilterForColor(m)
-	} else {
-		return FilterForColor(m)
-	}
-}
-
-func FilterForNode(m model) []k8s.Node {
-	var filteredNodes []k8s.Node
-	FilterNodeInput := strings.Split(m.args.FilterNodes, ",")
-
-	// Creating a new map to store the values of NodeStats list
-	// Choosing Map over Nested Array for comparision is best for TimeComplexity
-	//NodesMap := make(map[string]k8s.Node)
-
-	for _, node := range m.nodestats {
-		// NodesMap[node.Name] = node
-		for _, FilteredNode := range FilterNodeInput {
-			if matched, _ := regexp.MatchString(FilteredNode, node.Name); matched {
-				filteredNodes = append(filteredNodes, node)
-			}
-		}
-	}
-
-	if len(filteredNodes) > 0 {
-		utils.Logger.Debug("Filter For Node results", filteredNodes)
-		m.nodestats = filteredNodes
-		return m.nodestats
-	} else {
-		utils.Logger.Errorf("No matching Nodes found.. Exiting")
-		os.Exit(2)
-		return m.nodestats
-	}
-
-}
-
-func FilterForLabel(m model) []k8s.Node {
-	var filteredNodes []k8s.Node
-
-	FilterKey := strings.Split(m.args.FilterLabel, "=")[0]
-	FilterValue := strings.Split(m.args.FilterLabel, "=")[1]
-
-	if FilterKey == "" || FilterValue == "" {
-		utils.Logger.Errorf("Filter Key or Value is empty.. Exiting")
-		os.Exit(2)
-	}
-
-	for _, node := range m.nodestats {
-		if _, ok := node.Labels[FilterKey]; ok {
-			if node.Labels[FilterKey] == FilterValue {
-				filteredNodes = append(filteredNodes, node)
-			}
-		}
-	}
-
-	if len(filteredNodes) > 0 {
-		utils.Logger.Debug("Filter For Label results", filteredNodes)
-		m.nodestats = filteredNodes
-		return m.nodestats
-	} else {
-		utils.Logger.Errorf("No matching Nodes found.. Exiting")
-		os.Exit(2)
-		return m.nodestats
-	}
-}
-
-func FilterForColor(m model) []k8s.Node {
-	utils.Logger.Debug("Filter for Color called")
-	var filteredNodes []k8s.Node
-	var thresholdMin, thresholdMax float64
-
-	// Define the color threshold values
-	switch m.args.FilterColor {
-	case "red":
-		thresholdMin = 70
-		thresholdMax = 100
-	case "orange":
-		thresholdMin = 30
-		thresholdMax = 70
-	case "green":
-		thresholdMin = 0
-		thresholdMax = 30
-	default:
-		thresholdMin = 0
-		thresholdMax = 100
-	}
-
-	// Filter nodes based on metric and threshold values
-	for _, node := range m.nodestats {
-		var usagepercent float64
-		switch m.args.Metrics {
-		case "memory":
-			usagepercent = float64(node.Usage_memory_percent) / 100.0
-		case "cpu":
-			usagepercent = float64(node.Usage_cpu_percent) / 100.0
-		case "disk":
-			usagepercent = float64(node.Usage_disk_percent) / 100.0
-		default:
-			if m.args.Debug {
-				fmt.Println("No Matching Metric", m.args.Metrics)
-			}
-		}
-
-		if (usagepercent*100) >= thresholdMin && (usagepercent*100) < thresholdMax {
-			filteredNodes = append(filteredNodes, node)
-		}
-	}
-	if m.args.Debug {
-		fmt.Println("Filter For Color result:", filteredNodes)
-	}
-	return filteredNodes
-
-}
-
-func PrintDesign(output *strings.Builder, maxNameWidth int) {
-	lines := strings.Repeat("-", maxNameWidth+12+12+20)
-	fmt.Fprint(output, lines)
-	fmt.Fprint(output, "\n")
-}
-
-func getUnit(metricType string) string {
-	unit := ""
-	switch metricType{
-		case "memory": unit = "MB"
-		case "cpu": unit = "Cores"
-		case "disk": unit = "GB"
-	}
-	return unit
-}
-
-func headlinePrinter(m *model, output *strings.Builder, Nodes *[]k8s.Node, maxNameWidth *int) {
-
-	
-	unit := getUnit(m.args.Metrics)
-	freeHeading := "Free(" + unit+")"
-	maxHeading := "Max(" + unit+")"
-
-	values := []interface{}{"Name", freeHeading, maxHeading, "Pods"}
-	if m.args.LabelToDisplay != "" {
-		m.format = "%-" + strconv.Itoa(*maxNameWidth) + "s %-10s %-10s %-5s %-12s %s\n"
-		values = append(values, m.args.LabelAlias, "Usage%")
-		*maxNameWidth = *maxNameWidth+12
-	} else {
-		m.format = "%-" + strconv.Itoa(*maxNameWidth) + "s %-10s %-10s %-5s %s\n"
-		values = append(values, "Usage%")
-	}
-	fmt.Fprintf(output, m.format, values...)	
-	
-}
-
-func MetricsHandler(m model, output *strings.Builder) {
-
-	// Nodes Filtering based on filters
-	filteredNodes := ApplyFilters(m)
-
-	m.nodestats = filteredNodes
-	SortByHandler(m)
-
-	// decide formatting and Maximum width
-	maxNameWidth := 30
-	for _, node := range filteredNodes {
-		if maxNameWidth < len(node.Name) {
-			maxNameWidth = len(node.Name)
-		}
-	}
-	// Header and Version info
-	
-	fmt.Fprint(output, "\n# KubeNodeUsage\n# Version: 3.0.2\n# https://github.com/AKSarav/Kube-Node-Usage\n\n")
-
-	if !m.args.NoInfo {
-		fmt.Fprint(output, "\n# Context: ",m.clusterinfo.Context,"\n# Version: ",m.clusterinfo.Version,"\n# URL: ",m.clusterinfo.URL,"\n\n")
-	}
-
-	fmt.Fprint(output, "# ", strcase.ToCamel(m.args.Metrics)," Metrics\n\n")
-	headlinePrinter(&m ,output, &filteredNodes, &maxNameWidth)
-	PrintDesign(output, maxNameWidth)
-
-	if m.args.Metrics == "memory" {
-		for _, node := range filteredNodes {
-			prog := GetBar(float64(node.Usage_memory_percent) / 100.0)
-			values := []interface{}{node.Name, strconv.Itoa(node.Free_memory/1024), strconv.Itoa(node.Capacity_memory/1024), node.TotalPods}
-			if m.args.LabelToDisplay != "" {
-				values = append(values, node.LabelToDisplay, prog.ViewAs(float64(node.Usage_memory_percent)/100.0))
-			} else {
-				values = append(values, prog.ViewAs(float64(node.Usage_memory_percent)/100.0))
-			}
-			fmt.Fprintf(output, m.format, values...)
-		}
-	} else if m.args.Metrics == "cpu" {
-		for _, node := range filteredNodes {
-			prog := GetBar(float64(node.Usage_cpu_percent) / 100.0)
-			values := []interface{}{node.Name, strconv.Itoa(int(node.Free_cpu)), strconv.Itoa(node.Capacity_cpu), node.TotalPods}
-			if m.args.LabelToDisplay != "" {
-				values = append(values, node.LabelToDisplay, prog.ViewAs(float64(node.Usage_cpu_percent)/100.0))
-			} else {
-				values = append(values, prog.ViewAs(float64(node.Usage_cpu_percent)/100.0))
-			}
-			fmt.Fprintf(output, m.format, values...)
-		}
-	} else if m.args.Metrics == "disk" {
-		for _, node := range filteredNodes {
-			prog := GetBar(float64(node.Usage_disk_percent) / 100.0)
-			values := []interface{}{node.Name, strconv.Itoa(node.Free_disk/1024/1024), strconv.Itoa(node.Capacity_disk/1024/1024), node.TotalPods}
-			if m.args.LabelToDisplay != "" {
-				values = append(values, node.LabelToDisplay, prog.ViewAs(float64(node.Usage_disk_percent)/100.0))
-			} else {
-				values = append(values, prog.ViewAs(float64(node.Usage_disk_percent)/100.0))
-			}
-			fmt.Fprintf(output, m.format, values...)
-		}
-
-	}
-}
 
 func checkinputs(args *utils.Inputs) {
 
@@ -415,6 +124,7 @@ func main() {
 		label string
 		lblAlias string
 		noinfo bool
+		pods bool
 	)
 
 	flag.BoolVar(&helpFlag, "help", false, "to display help")
@@ -427,6 +137,7 @@ func main() {
 	flag.StringVar(&metrics, "metrics", "memory", "choose which metrics to display (memory, cpu, disk)")
 	flag.StringVar(&label, "label", "", "choose which label to display")
 	flag.BoolVar(&noinfo, "noinfo", false, "disable printing of cluster info")
+	flag.BoolVar(&pods, "pods", false, "enable pod details")
 
 	flag.Parse()
 
@@ -462,6 +173,7 @@ func main() {
 		LabelToDisplay: label,
 		LabelAlias: lblAlias,
 		NoInfo: noinfo,
+		Pods: pods,
 	}
 
 	checkinputs(&args) // sending the args using Address of Operator
@@ -470,97 +182,28 @@ func main() {
 		PrintArgs(args)
 	}
 
-	// Model Intialized here - Start of the Program
-	mdl := model{}
-	mdl.args = &args
-	mdl.clusterinfo = k8s.ClusterInfo()
-	mdl.nodestats = k8s.Nodes(&args)
-	
-	if _, err := tea.NewProgram(mdl).Run(); err != nil {
-		fmt.Println("Oh no!", err)
-		os.Exit(1)
-	}
-}
-
-// tickMsg is the message we'll send to the update loop every second.
-type tickMsg time.Time
-
-// model is the Bubble Tea model.
-type model struct {
-	clusterinfo k8s.Cluster
-	nodestats []k8s.Node
-	args      *utils.Inputs
-	format	string
-}
-
-// Init Bubble Tea model
-func (m model) Init() tea.Cmd {
-	return tickCmd()
-}
-
-// Update method for Bubble Tea - for constant update loop
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		if msg.Type == tea.KeyCtrlC {
-			fmt.Println("Ctrl+C pressed")
-			return m, tea.Quit
-		}
-		//  check if Q or q is pressed
-		if msg.Type == tea.KeyRunes && (msg.Runes[0] == 'Q' || msg.Runes[0] == 'q') {
-			fmt.Println("Q or q pressed")
-			return m, tea.Quit
-		}
-
-		// // check if R or R is pressed
-		// if msg.Type == tea.KeyRunes && (msg.Runes[0] == 'R' || msg.Runes[0] == 'r') {
-		// 	fmt.Println("R or r pressed")
-		// 	m.nodestats = k8s.Nodes(m.args.Metrics)
-		// 	return m, tea.ClearScreen
+	if pods {
+		fmt.Println("Yet to implement")
+		// KubePodUsage - For Pods
+		// args.Pods = true
+		// mdl := podusage{}
+		// mdl.args = &args
+		// mdl.clusterinfo = k8s.ClusterInfo()
+		// mdl.podstats = k8s.Pods(&args)
+		// if _, err := tea.NewProgram(mdl).Run(); err != nil {
+		// 	fmt.Println("Oh no!", err)
+		// 	os.Exit(1)
 		// }
-	case tickMsg:
-		m.clusterinfo = k8s.ClusterInfo()
-		m.nodestats = k8s.Nodes(m.args)
-		return m, tea.Batch(tickCmd())
+	} else{
+		// KubeNodeUsage - For Nodes
+		mdl := nodemodel.NodeUsage{}
+		mdl.Args = &args
+		mdl.ClusterInfo = k8s.ClusterInfo()
+		mdl.Nodestats = k8s.Nodes(&args)
+		
+		if _, err := tea.NewProgram(mdl).Run(); err != nil {
+			fmt.Println("Oh no!", err)
+			os.Exit(1)
+		}
 	}
-	return m, nil
-
-}
-
-func GetBar(decider float64) progress.Model {
-
-	decider = decider * 100
-
-	var prog progress.Model
-	// decide which color to use based on the usage percentage below 30% is green, above 70% is red, else yellow
-	if decider < 30 {
-		prog = progress.New(progress.WithScaledGradient("#0bad5d", "#74b03f"))
-	} else if decider > 70 {
-		prog = progress.New(progress.WithScaledGradient("#13B013", "#F11658"))
-	} else {
-		prog = progress.New(progress.WithScaledGradient("#13B013", "#F18016"))
-	}
-	return prog
-}
-
-// View renders bubble tea
-func (m model) View() string {
-
-	var output strings.Builder
-
-	DebugView(m, &output) // If debug on this would print Node and arg details
-
-	MetricsHandler(m, &output)
-
-	output.WriteString("\n" + helpStyle("Press Q or Ctrl+C to quit"))
-
-	return output.String()
-
-}
-
-// tickCmd returns a command that sends a tick every second.
-func tickCmd() tea.Cmd {
-	return tea.Tick(time.Second*1, func(t time.Time) tea.Msg {
-		return tickMsg(t)
-	})
 }
